+++ packages/terra-compact-interactive-list/src/proptypes/columnShape.js
@@ -1,26 +1,53 @@
 import PropTypes from 'prop-types';
+import { alignTypes } from '../utils/constants';
 
 const columnShape = PropTypes.shape({
   /**
    * Required string representing a unique identifier for the column.
    */
   id: PropTypes.string.isRequired,
+
   /**
-   * String of text to render within the column header cell.
+   * String of text which will be used to provide a column context and to serve as a column header if needed.
    */
   displayName: PropTypes.string,
+
   /**
-   * Number that specifies the minimum column width in pixels.
+   * A valid css string, px, em, or rem supported (should be the same across all width units).
+   * If not set, the column will be considered a flex growing column.
+   * If width is set, but flexGrow prop is also set to true, the width will be disregarded.
    */
-  minimumWidth: PropTypes.number,
+  width: PropTypes.string,
+
   /**
-   * Number that specifies the maximum column width in pixels.
+   * A valid css string, px, em, or rem supported (should be the same across all width units).
+   * Will be disregarded if the width is set and flexGrow is omitted or false.
+   * Is reccomended in cases the column is flex growing column.
    */
-  maximumWidth: PropTypes.number,
+  minimumWidth: PropTypes.string,
+
   /**
-   * A number (in px) specifying the width of the column. If not provided, the Table's default column width will be used.
+   * A valid css string, px, em, or rem supported (should be the same across all width units).
+   * Will be disregarded if the width is set and flexGrow is omitted or false.
+   * Is reccomended in cases the column is flex growing column.
    */
-  width: PropTypes.number,
+  maximumWidth: PropTypes.string,
+
+  /**
+   * Whether the column width can grow (a flex growing column).
+   * Defaults to false if the width is set and true if the width is not set.
+   */
+  flexGrow: PropTypes.bool,
+
+  /**
+   * Aligns the cell content to center, righ, or left. Defaults to the left.
+   */
+  align: PropTypes.oneOf([
+    alignTypes.CENTER,
+    alignTypes.RIGHT,
+    alignTypes.LEFT,
+  ]),
+
 });
 
 export default columnShape;