+++ packages/terra-framework-docs/src/terra-dev-site/test/compact-interactive-list/DefaultCompactInteractiveList.test.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import CompactInteractiveList from 'terra-compact-interactive-list';
-
-const DefaultCompactInteractiveList = () => (
-  <CompactInteractiveList />
-);
-
-export default DefaultCompactInteractiveList;