+++ packages/terra-framework-docs/CHANGELOG.md
@@ -9,6 +9,7 @@
   * Added new test for Terra Slider component for long field labels.
   * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
   * Added examples and tests to `terra-data-grid` for sections in FlowsheetDataGrid.
+  * Added documentation and examples for `terra-compact-interactive-list`.
   
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.